Move configuring the AcmeClient to AcmeService Move PromptYesNo to a method, makes a bit more sense
