+++ letsencrypt-win-simple/Extensions/StringExtensions.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 
 namespace LetsEncrypt.ACME.Simple.Extensions
@@ -10,18 +9,5 @@ namespace LetsEncrypt.ACME.Simple.Extensions
             =>
                 Path.GetInvalidFileNameChars()
                     .Aggregate(fileName, (current, c) => current.Replace(c.ToString(), string.Empty));
-        
-        public static bool PromptYesNo(this string message)
-        {
-            while (true)
-            {
-                var response = Console.ReadKey(true);
-                if (response.Key == ConsoleKey.Y)
-                    return true;
-                if (response.Key == ConsoleKey.N)
-                    return false;
-                Console.WriteLine(message + " (y/n)");
-            }
-        }
     }
 }
